whatsapp-supabase/requirements.txt
supabase==2.4.5
httpx==0.27.0
tenacity==9.0.0
python-dotenv==1.0.1
